Move and pass notebook test
