+++ sphinx_gallery/gen_rst.py
@@ -674,7 +674,7 @@ def generate_file_rst(fname, target_dir, src_dir, gallery_conf):
         example_rst += "**Total running time of the script:**" \
                        " ({0: .0f} minutes {1: .3f} seconds)\n\n".format(
                            time_m, time_s)
-        example_rst += CODE_DOWNLOAD.format(fname, example_nb)
+        example_rst += CODE_DOWNLOAD.format(fname, example_nb.save_file())
         example_rst += SPHX_GLR_SIG
         f.write(example_rst)
